Make installed capacity the same in all examples
Some examples had installed capacity of 10 W instead of 6 MW.
For gen_commit_bin/lin generators, who use this as the de facto unit
size, this is not consistent with the heat rate curves inputs which
assume a unit size of 6 MW and as a result it made the fuel burn much
larger than expected, due to the intercept term being much larger (it
is multiplied by the number of units online, which is affected by the
unit size; note that the slope is unaffected).

The unittests are updated so that they still pass.