Add spinup_or_lookahead to timepoint index
This would allow us to reuse a timepoint ID that is part of a
spinup/lookahead in one horizon and a 'regular' timepoint in another
horizon. It would mean that the spinup_or_lookahead would have to be
either 0 or 1, not NULL. The results queries are updated accordingly.

Example:
The usual thing we do might be to run a full year as the subproblem
with days as a horizon (no overlaps). That would create seams across
the horizons whether they are circular or linear. So, instead, we could
run the year, but have each horizon actually consist of the day we're
interested in and the previous day as spinup to mitigate the seams.
That would mean that day 2 would have day 1 as spinup (let's call that
horizon 2) and day 3 would have day 2 as spinup (let's call that
horizon 3). Horizon 2 an 3 in this case have the day 2 timepoints
overlapping.

Note: having overlapping timepoints between horizons might break the
dispatch plots because the results_project_dispatch_by_technology table
does not have a horizon column (since there is no balancing type at the
technology level), and we just get the timepoints based on horizon from
the results_project_dispatch and match to that, but then find some of
these timepoints more than once in the
results_project_dispatch_by_technology, as they also exist in a previous
subproblem/horizon. A seperate PR will address this issue and replace
the horizon as an input to the dispatch plot with a starting and
ending timepoint.

Closes #472