Make sure results plot exclude spinup/lookahead (#507)
Only do so when appropriate. For the carbon and RPS plots, the spinup
and lookahead timepoints are counted towards the target/cap so they
should be included in the results. Generally we don't expect to have
spinup/lookahead timepoints when running a carbon cap or RPS target
scenario.

Also, the results summary.txt file includes the spinup and lookahead 
timepoints when calculating the total energy. I added a comment in 
the code to make this clear.

Closes #486