Add carbon emissions to fuel burn results
Carbon emissions are not output unless you were running a scenario
with a carbon cap. It is likely though that users want to look at
emissions without running a scenario with a cap, e.g. when doing
production simulation. You could get around this by adding a really
large cap but that still requires users to make up inputs for the carbon
cap zones and assigning carbon cap zones to projects.

With this update, the fuel burn table now includes the carbon emissions
as well. It doesn't require any additional user inputs since the
emission factors were already part of the fuels table.

The down side is that there are now duplicate inputs between the fuel
burn table and the carbon emissions table, and there could be a
discrepancy between the tables if there are projects that don't have
a carbon cap zone. This could be fixed by removing the carbon emissions
table entirely and simply using the fuel burn table. However, that
would remove some of the analogy between the carbon module and rps
modules (right now they're pretty much the same structure).