Define OP_TMPS for var op types in capacity package (#523)
In general, we create the TYPE_OPERATIONAL_TIMEPOINT sets from the
PROJECT_OPERATIONAL_TIMEPOINTS (based on definitions in the capacity
package) within each operational module. We were overwriting that
behavior in the two variable operational types by loading data for the
set from the capacity factor inputs. Operational status should be
defined via the capacity type and associated inputs, and the operations
inputs should conform to that, so the overwriting behavior is removed
here.