Variable O&M by loading level (#510)
This PR adds a second component to the variable O&M that varies with the loading level. This component is additional to the already existing variable O&M, and is only available for operational types with commitment variables (gen_commit_lin/bin, gen_commit_cap) and gen_always_on.

The input data comes in through a new input table variable_om_curves, which is loaded in the project.init module. The model components (variables and sets) are created in project.init as well. Operational modules that use this functionality have an extra set, decision variable, and constraint, and use this in combination with the components read in at project.init to govern this new variable O&M cost. This exactly mirrors the approach on how heat_rate_curves are input/loaded and constrained.

All operational modules now have a variable_om_cost rule that return the total variable O&M cost (again, similar to the heat rates/fuel burn). For operational types without this new functionality, this will be simply the old variable O&M cost. For operational types with this new functionality, it will consist of 2 components: the old variable O&M cost and the variable O&M by loading level. The second (new) component can be easily ignored by simply leaving the variable_om_cost_scenario_id blank for the project in question in the operational characteristics table).

Closes #429