Initial refactoring of project inputs (#528)
The project.__init__ module now defines, gets from the database, and
loads only the params required for all projects, i.e. the load zone an
"types", as well as the optional technology, which is only used on the
output side. It creates the projects.tab file.

We now get the rest of the operational characteristics from the
database and write them to an input file in the
project.operations.__init__ module. This is done by appending to the
projects.tab input file. The function for doing this is factored out 
in project.common_functions. The variable_om_cost_per_mwh parameter is
now also defined and data for it loaded in project.operations.__init__
(along with the fuels-related params).