Refactor gen_commit_cap load_module_specific_data (#531)
We create a CSV that contains the data for whether specific
operational characteristics are required or optional for the
gen_commit_cap module. Based on that, there is now a function named
get_optype_inputs_as_df in operational_types.common_functions that
creates a sub-dataframe from the projects.tab file with just the
relevant rows and columns for the operational type.

There is another function called get_param_dict in
operational_types.common_functions that creates the param dictionary
to be loaded into Pyomo from the sub-dataframe and the relevant column
name.

In the load_module_specific_data method of the operational type module,
we now simply iterate over the required and optional columns, and use
the get_param_dict function to load the relevant data. Being able to
simply loop like this is dependent on the param name conforming to
certain requirements: it must consist of the operational type name, an
underscore, and the column name (this is true in the gen_commit_cap
module).