Fix scenario_description.csv LF endings (#532)
Make sure we change the csv.write() to write out LF endings instead of
the default CRLF endings. That way all files that are tracked are LF
so we don't get the issue where the locally written CRLF version is
converted to LF upon committing to the repo, and next time you write
that file again locally with CRLF line endings, you see a diff that
shouldn't be there.

Note: the .csv files in the results which aren't tracked still have
the default settings for csv.write(), i.e. CRLF line endings. Also some
of the csv files that are created manually in db/data and
db/csvs_test_examples have CRLF endings, or even CR endings. The latter
shows up when I save csv files on my Mac for Excel version (a known
bug). That's probably okay because we don't touch those files except
when we change them (compared to the .tab and .csv file in the tests
that are re-created every time you run unittests).

Closes #415