Add input validation to test_examples unittests
First step is to add it to all test examples and make sure the input
validations pass.
The second step could be to make sure that no input examples have
any input validation violations (I am pretty sure they have some now).

WIP to check whether this is the place we want to do this and whether
we think it's okay to do this for every single test example.