Refactor subproblems and stages IO (#544)
This is a refactor of how we deal with determining the subproblem and
stage structure of a scenario to make it less convoluted.

The process is now as follows:

The number of subproblems and stages per subproblem is determined
directly from the database inputs as before (see SubProblems class in
scenario_chars.py).

Based on that, we create the appropriate directory structure and write
out all the .tab files for the scenario in get_scenario_inputs.py.

The subproblems.csv files are not written anymore in
get_scenario_inputs.py nor used anywhere. Instead of using those files,
in run_scenario.py, we determine the correct problem to solve entirely
based on the directory structure. The ScenarioStructure class is
refactored accordingly.

Subscenario and stage directories must be integers and figuring out
what those are is factored out in
gridpath.auxiliary.check_of_integer_subdirectories as we have to do
that in a number of places.

The of_multi_stage feature is removed entirely. Instead, the
determine_modules function figures out whether to remove the modules
previously associated with the multi-stage feature:
1) in get_scenario_inputs.py, we determine whether any of the
subproblems have stages based on the SubProblems structure for the
scenario (i.e. based on the database inputs); if they do, we pass a
flag to the determine_modules function to ensure that the modules
previous associated with the multi-stage feature are included;
2) in run_scenario.py, we simply pass the scenario directory to
determine_modules (as well as None for the features_list as before and
None for the stages_flag), with the function now determining whether
there are stages (meaning include the fix_commitment module) based on
the scenario directory structure (note that the function still looks
for the features.csv file to determine what other modules to use).

In summary, this removes the need for the multi-stage optional feature
and for the subproblems.csv files. Instead, when getting inputs, we
determine the directory structure and whether to call the
fix_commitment module based on the database temporal inputs and then,
when running a scenario, use the directory structure to solve the
appropriate subproblem/stage problems and decide whether to call the
fix_commitment module at each subsequent run_e2e step.