Pass the scenario directory on writing model inputs (#546)
We were passing the final inputs directory (that depends on subproblem
and stage) and what the scenario directory is would need to be reverse-
engineered every time in write_model_inputs. Instead, we now pass the
scenario directory directly instead and determine the inputs directory
based on the subproblem and stage arguments, which are already being
passed to write_model_inputs (this also re-standardizes with the
treatment in the other module methods, which figure out the correct
inputs directory).

The "reverse engineering" now happens in the
get_inputs_from_database functions, which use 1 for the subproblem and
stage whenever those are empty strings.