Refactor loading in variable profiles (#543)
This PR refactors loading the inputs, getting the inputs from the database, and writing the inputs to .tab files. It moves the common code between gen_var and gen_var_must_take to the common functions to avoid duplicate code across the different operational types. 

This also introduces more dataframe operations to do filtering, which
presumably (not tested) is more efficient and could be easier to follow.