Refactor writing tab files
For now just refactor hydro, var profiles and startup_chars, which
already had this function re-factored into common inputs.

In a second step, we can introduce this function in all modules that
write tab files. However, before we do that we need to make sure that
the tab file column names and db columns names are the same, since this
function derives the tab file column names from the db columns, using
the query results description.