Fix zero division error
When the availability is zero or the min stable level is zero, Pmin
will be zero, and dividing by Pmin will result in a division by zero
error.

This will fix this but still leaves open the issue what happens if you
mix gen_commit_lin/bin op-types with gen_new_lin/bin op-types. I did
a quick test example to check this using the ipopt non-linear solver
and didn't get an error, but the results looked off with synced units
being a non-integer number, e.g. 1.5.

We could simply replace the capacity with a very big number but we need
to make sure that number is larger than Pmin.