Refactor write inputs (#556)
For now just refactor hydro, var profiles and startup_chars, which
already had this function re-factored into common inputs.

This required aligning the hydro column names between the .tab
files and the database tables (they weren't), since the .tab file
columns are now derived from the database query results.

In a second step, we could introduce this function in all modules that
write tab files. However, before we do that we need to make sure that
the tab file column names and db columns names are the same
everywhere (they are not aligned everywhere); low priority but noting
this here for posterity.