Fix broken validation script (#557)
With #544, we introduced the need to pass a flag to determine whether
the stage-related modules should be included when getting inputs (the
value of the flag is determined based on the scenario's temporal
structure). The same flag needs to be passed in validate_inputs.py, as 
that script also works directly off the database without looking at the
directory structure and files. Without passing this flag, the validation
script would break.