Support for specifying RPS targets as % of load (#562)
This implements the ability to specify RPS target as percent of the 
total period load in some pre-specified load areas. This target can be 
specified instead of or in addition to the energy target we already had
in place. If not specified, both the energy target and the percentage
targets will now default to 0. The two components are summed in the new
RPS_Target expression that is now used instead of rps_target_mwh in the
RPS balance constraint.

To use this feature, the user must specify a percentage for each RPS
zone and period, and a mapping from the RPS zone to the load zones 
whose load should be part of the calculation (all part of the same 
rps_target_scenario_id, although contained in two different tables).
The former is loaded via rps_target.tab and the latter via a new input
file, rps_target_load_zone_map.tab, which is only written if the user 
specifies a mapping.

If the user does not specify a RPS zone to load zones mapping, the 
percentage-based component of the RPS target will be 0. This is done 
by initializing the RPS_ZONE_LOAD_ZONES map set as empty list when not 
finding the map input file and checking for that condition in 
rps_target_rule.

I added an example that has the same RPS target as the 
"2periods_new_build_rps" example, but specified as percent of load. 
The objective functions of the two are the same.

Note that now that the RHS of the RPS balance equation is an 
expression, the sign of the RPS constraint dual changed (see the 
changes in summary_results.txt of the RPS examples). I don't have a 
complete explanation for it, but the absolute values of the duals are 
the same, so I think it's just how the solver interprets a param vs an 
expression.