Support for linked subproblems (#553)
This implements the ability to link the results from one subproblem
as inputs to the next subproblem. This is done for the operational
modules only, i.e. the ramp, min up/down time, and startup trajectory
functionality.

The inputs_temporal_timepoints table now includes an additional column
called linked_timepoint. The user can designate timepoints from the
last horizon of a subproblem to be linked to the next subproblem. These
linked timepoints should be non-positive integers, ordered, and the
most recent linked timepoint should be 0. If linked timepoints are
specified for a subproblem, GP will use those as the previous
timepoints for the first horizon of the next subproblem
(subproblem_id + 1) only if the first horizon of the next subproblem
has a 'linked' boundary, a new horizon boundary type implemented here.

When getting inputs, GP now looks for linked timepoints when getting
the scenario inputs (see write_linked_subproblems_map in
get_scenario_inputs.py) and a writes a map file to the scenario
directory if it finds any and not otherwise. Whether or not this map
file is written is the flag we use later on to decide whether to
initialize the LINKED_TMPS set, and the hrs_in_linked_tmp and
furthest_linked_tmp params in timepoints.py and whether to export
results to link (for the latter see the check_for_tmps_to_link
function in
gridpath/project/operations/operational_types/common_functions.py,
which is used when op type modules export results). Each op type module
is aware of which results it needs to export to be loaded into the next
subproblem and, if it finds the linked timepoints map CSV file, writes
these to tab files in the linked subproblem's input directory, applying
the mapping. The op type modules look for these input files when
loading their inputs and simply pass if they don't find them, not
initializing the linked components.

The formulation of the operational type modules needed to be changed to
account for what may have happened in the linked timepoints. For the
purpose the 'determine_relevant_timepoints' function in
gridpath/project/operations/operational_types/common_functions.py had
to be adapted to also return the relevant linked timepoints in addition
to the relevant timepoints within the horizon.

Note that here and in the op type modules itself, we rely on the fact
that the linked timepoints are ordered, non-negative integers, and
assume the most recent one has index 0. In the op type modules, when we
only need to look at the previous timepoint for the first timepoint of
a "linked" horizon, we simply take this 0 index.

Finally, the formulation of constraints looking at previous timepoints
in the operational type modules had to be updated to account for the
cases when the linked timepoints need to be considered. The goal was to
not "change" the formulation per se, but stay within our formulation
framework, and just add if logic within each constraint to apply the
linked timepoint inputs (instead of/in addition to variables) when
needed.