Standardize carbon emission units
Rather than using a mixture of "tonnes" and "MMT", and converting
between the two in the code base as appropriate, we should use the same
unit everywhere so that ultimately users can choose what unit they use,
without causing conversion issues in the code base.

This commit ensures the same units are used everywhere. This is a
pre-requisite for the flexible units update (#565).