Support for specifying reserves as % of load (#567)
This implements the ability to specify all reserves as percent of the 
load in some of the LOAD_ZONES in each timepoint. This in addition to 
the by-timepoint specification of reserves we already had in place: 
the two are additive in a new expression component, which is now used 
in the reserve-balance constraints, and either specification can be 
omitted, defaulting to 0.

The CSV-to-DB structure had to change some as each 
reserve_type_scenario_id now consists of several inputs: the 
by-timepoint specification, the percent specification, and the 
BA-to-LZs map for the percent specification. We therefore now use our 
subscenario directory parsing tools.

To use this feature, the user must specify a percentage for each 
reserve BA, and a mapping from the BA to the load zones whose load 
should be part of the calculation (all part of the same 
reserve_scenario_id, although contained in two different tables). They 
are loaded via two new input files, one specifying the percent 
requirement and the other one the mapping. Neither of those files is 
written if a mapping is not specified for this reserve_scenario_id. 
In that case, the percent requirement defaults to 0 and the an empty 
list is loaded for the map set.

I added an example that is based on the 'test' example but also has a 
percent requirement specified for load following down in addition to 
the by-timepoint requirement.

Note that, wherever it was not already done, I refactored the various 
reserve modules to use the same functions for the IO workflow in 
implementing this new functionality. I did also change the name of the 
reserve_type_requirement.tab file to also include _tmp in order to 
distinguish from from the new reserve_type_percent_requirement.tab 
file, which does affect a number of files in the examples.

Also note that the by-timepoint requirement is now specified for all 
BA-timepoint combinations, not for a reserve-specific set based on the 
by-timepoint requirement input file. This is because we now have the 
parameter default to 0 if data are not loaded for a particular BA-tmp 
combination.