Introduce Flexible Units (#565)
This update introduces a new table, `mod_units`, which specifies the units for each metric used in the model. The table contains "core" units, which can be set by the user, and "secondary" units, which are combinations of core units.

Upon database creation, if the `--custom_units flag` is included, a prompt asks the user to specify the core units. Next, the secondary units are derived from the core units. This unit database table is used in the viz plotting scripts to set the appropriate axis and hover tool labels. The table could also be exposed in the UI to guide users on the units that are used.

Get_scenario_inputs now also creates a `units.csv` file which specifies the units used in the scenario (based on the new mod_units table) so that non-database dependent modules have access to this information. This `units.csv` file is used to flexibly write out the appropriate units in the summary_results.txt file.

Note: this PR does not update any of the variable name or headers, which have units hard-coded into them. Once units are flexible, those hard-coded units are no longer (always) valid, so they will need to be updated (see next steps).

Next steps:
- remove units from all database table column names and csv/tab file headers. Note: might want to keep some type of unit designation, perhaps generalized, to help users understand what's in the column (e.g. `heat_rate_mmbtu_per_mwh` --> `heat_rate_fuel_energy_per_electric_energy`)
- remove units from all variable names in the code (same note as above).
- expose table to UI?