Fix bounds for linked subproblem inputs (#577)
Depending on tolerance settings, the values of variables exported from 
one subproblem to be linked as params to the next subproblem can be 
slightly out of bounds (e.g. just above 1 for Binary or 
PercentFraction, very slightly negative for NonNegativeReals, etc.). 
This applies the correct bounds on exporting the results to avoid 
throwing a `Value not in parameter domain` error when loading the 
linked inputs for the next subproblem.