Clear all prior results before import (#583)
This makes sure that all prior results for a scenario_id are deleted 
before importing new results. While all modules currently delete their 
own results before importing again, this addition takes care of the 
situation in which modules were changed for a scenario and perhaps not 
included in a subsequent run, which could result in phantom results in 
some tables.

Note that if using the UI, the user is required to clear all results 
and statuses before re-running a scenario out of abundance of caution, 
but this addition helps those running scenarios from the command line.