Capacity constraints for groups of projects (#581)
This implements the ability to impose constraints on the capacity 
(minimum and maximum new build, and minimum and maximum total capacity)
of groups of projects in a given period.

For the purpose, we have two new subscenario IDs specifying the 
capacity group requirements and group-to-project mapping as well as a 
new GP module with the constraint formulation. This new module checks 
if the subscenarios are specified and writes the input files if so; 
otherwise, the input files are not written, data not loaded, the 
components are initialized with default values (non-binding), and 
results are not exported from the model and not imported into the 
database.

Note that we haven't added the ability to view/select/edit these new 
subscenarios in the UI due to an unrelated issue with the number of 
joins in the scenarios_view we use. Once I come up with a workaround 
for that, I will add the ability to select these subscenarios from the 
UI.