Track solver termination condition (#592)
This implements tracking of the solver status and changes some 
behavior according to the status.

When the problem is infeasible, the error is often rather obscure, as 
it usually involves an uninitialized variable error that does not get 
thrown until the results-export step. We now check the solver status 
and, if not optimal, print a warning and stop the results-export, 
summarize-results, and db-import steps for the non-optimal 
subproblem/stage. In a linked subproblem setting, we stop the run as 
we don't have results to link as inputs to the next subproblem; 
otherwise, we can continue to the next subproblem if there is one.