Align min_stable_level naming
The minimum stable level was called "min_stable_level" in the database
column but within the model, and in the .tab files it was called
"min_stable_level_fraction", which results in issues when we try to
refactor input validation and data loading.

This update aligns it so it is called "min_stable_level_fraction"
everywhere.