Fix NumPy FutureWarning
Before, when running the validation I would get the following error
for test example with scenario ID 47:
```
FutureWarning: elementwise comparison failed; returning scalar,
but in the future will perform elementwise comparison
```

I found out it was caused by the gen_must_run column which doesn't
have any entries. That column will therefore not be a string, and
the boolean comparison df['op_type'] == 'op_type' results in the
FutureWarning.

Forcing all columns to be strings seems to fix this issue.

See here for more info:
https://stackoverflow.com/questions/40659212/futurewarning-elementwise-
comparison-failed-returning-scalar-but-in-the-futur