Consolidate database utilities (#605)
This further refactors the port_csvs_to_gridpath.py script and removes 
all modules under csvs_to_db_utilities. The utilities data insertion 
functions are modified to load in lists of tuples. We now have a 
generic function that reads in CSVs and converts them to lists of 
tuples, then passes them to the utilities, thus removing the need for 
other data manipulation. Subscenarios contained in a single file now 
use this method. Note that the subscenarios requiring multiple files 
still use a separate method that was already in place (see 
db.utilities.common_functions) and the operational characterists are 
contained in a "wide" file, so also rely on a different method (i.e. 
not on column order but on matching CSV and database table column 
names).