Refactor write_validation_to_database function (#608)
When we were writing out validation results to the status_validation table, we had to construct a list of validation errors, where each validation error was a list of tuples, and then feed this to the write_validation_to_database() function. To make this process a little less verbose, and to make it more explicit what we are passing to the writing function, I refactored the function to take in the list of validation errors, and the associated meta-data (what module, what database table, etc.) and have it create the database rows inside the function. If there are no errors (empty list), the function does nothing.

I also simplified the status_validation schema somewhat by removing redundant columns (subscenario, and issue_type) and shortening the column headers.