Consolidate subscenario/inputs insert functions (#613)
This refactors the CSV-to-DB loading of all "simple" subscenarios
(i.e. subscenarios contained in a single file) into a single function,
with a number of the database "utilities" now completely unnecessary
and therefore deleted. The new function is called
generic_insert_subscenario in db.utilities.common_functions.

Instead of calling each utility's insert function separately in
port_csvs_to_gridpath.py, we now iterate over the rows in the CSV
master file to determine which subscenarios can use the
generic_insert_subscenario and use that instead. For this purpose,
the CSV master file has two new columns: the "subscenario_type" which
can be "simple," i.e. using the method implemented here, or "custom,"
i.e. still using its preexisting method, and the "project_input" to 
determine which whether this is a project-level subscenario (e.g. 
hydro chars, variable profiles, heat rates, etc.).

All changes to the schema and CSV files were required to properly 
align the CSVs and database tables, and get this to not throw any 
errors.