Refactor loading of multi-file subscenarios (#614)
This consolidates how directory-based (i.e. multi-file to multi-table) 
subscenarios are loaded from CSVs to the database. This applies to RPS 
targets, reserve requirements, and the ELCC surface, and their 
respective modules from db.utilities are now deleted. The required 
files for each subscenario are now listed in the CSV master file. The 
subscenario info is loaded once along with the respective input table 
by specifying the dir_main flag in the subscenario_type column; the 
rest of the files are loaded with the dir_aux flag.