Use standard method for loading opchar from CSV (#616)
Load operating characteristics from CSV via the standard method. 
Adding the auxiliary consumption field to a few of the test example 
CSVs was required to match the CSV and table headers. The 
db/utilities/project_operational_chars.py is not necessary anymore 
and therefore deleted. The inputs_project_all table is now also 
defunct so is deleted from the schema.