Refactor db.utilities.common_functions (#620)
This is a refactor of db.utilities.common_functions. The main changes
are:

The csv_read_data and csv_read_project_data are consolidated.

There are separate functions for loading single subscenarios (from
both a single CSV or a multi-CSV directory) and functions for
loading all subscenarios in a directory (with the looping done within
those functions).

The port script loops over the directories specified in the CSV master 
file and uses these latter functions to load all subscenarios within 
the directory. The idea with these changes is that both the 
load-single-subscenario functions and the 
load-all-subscenarios-in-directory functions can be called separately 
outside of a loop for greater flexibility in loading data.