Use standard methods for temporal CSV-to-DB (#622)
The utility to load temporal subscenarios are now refactored to take 
advantage of the common methods in db.utilities.common_functions. 
The CSV-to-DB functionality now also includes the ability to run a 
"custom method" -- used here to populate some of the temporal tables 
based on the data in other temporal tables (as opposed to having to 
duplicate data in the CSVs).