Add hours per period adjustment
When running multiple subproblems within one period, we need to adjust
the annualized capacity costs down according to the ratio of the number
of hours in the subproblem vs the number of hours in the period.

Here we add this adjustment through and additional input param that
specifies how many hours there should be in the period (must be 1 year
so 8760-8784) and a derived param that calculated the number of hours
in a given subproblem (taking into account the timepoint weights).

We also add a database view that implements this adjustment when
calculating the capacity costs for a scenario-period-stage-loadzone.