Scale down capacity cost expressions
So now capacity costs in the model (and in results) are scaled
down for the subproblem length. This gets rid of the need of a view
that scales down these capacity costs after the fact.

Also add this scaling down, plus reporting of the number of hours in
the subproblem period, for the transmission capacity costs.

To pass unittests we will first need to change the unittests timepoint
durations of the examples so that they are a full year's worth.