Update timepoint weights in examples
Make sure all periods (can be broken out in subproblems) represent 8760
hours. This is a prerequisite for the update that adds
the hours_in_full_period and adjusts the capacity costs in each period
depending on how many hours are represented in the period. In that
update we will validate that across all subproblems, the hours
represented should be equal to the hours in a full period. This is
not the case in the examples, and we will get a validation error,
which will make the unittests fail.

Note that a lot of the test examples have unserved energy, and adding
timepoint weights will make the operational costs, including
unserved energy orders of magnitude larger, which in turn can change
the dispatch decisions of the examples. We see this for the
examples 2periods_new_build_rps_w_rps_eligible_storage vs.
2periods_new_build_rps_w_rps_ineligible_storage and
2periods_new_build_rps_variable_reserves_subhourly_adj vs.
2periods_new_build_rps_variable_reserves which now have the
same objective function.