Add hours per period and relevant adjustments to capacity costs (#631)
When running multiple subproblems within one period, we need to adjust
the annualized capacity costs down according to the ratio of the number
of hours in the subproblem vs. the number of hours in the period.

Here we add this adjustment through an additional input param that
specifies how many hours there should be in the period (must be 1 year,
i.e. 8760-8784) and a derived param that calculated the number of hours
in a given subproblem (taking into account the timepoint weights and
durations, and excluding the spinup_or_lookahead timepoints).
Note that this required adding the spinup_or_lookahead info to the
.tab files and the model (new param), since before this was only used
on the database back-end.

Using these new params, capacity costs in the model (and in results) are
scaled down for the subproblem length. So when there are daily
subproblems, capacity costs in each subproblem (both the costs in
the objective function and the ones reported in the .csv files) are only
1/365th of the annualized capacity costs.

We also add validation to check that the hours represented in a period
across all subproblems is equal to the hours_in_full_period for all
periods and stagse.