Validate var profiles
Make sure that var profiles contain inputs for all project-timepoints.

We change the INNER JOIN to a LEFT JOIN so we can do similar checks
to what was done in #634 using the validate_missing_inputs function.
Note that the model will fail (error) with both the INNER JOIN and
LEFT JOIN if there are missing project-timepoints.