Hydro opchars, varprofile and avl validation (#638)
 - Make sure that hydro opchars contains inputs for all project-horizons.
 - Make sure that var profiles contain inputs for all project-timepoints.
 - Make sure that exogenous availability derates contain inputs for all 
   project-timepoints.

We change the INNER JOIN to a LEFT JOIN so we can do similar checks
to what was done in #634 using the validate_missing_inputs function.
Note that the model will fail (error) with both the INNER JOIN and
LEFT JOIN if there are missing project-horizons or project-timepoints
for the hydro opchars and var profiles. For the availibiltiy derates,
I've added a replace_nulls type function so the model still works if
there are missing inputs, even with the LEFT JOIN, but there is a
now a low severity validation error if that happens.

We also update validate_missing_inputs so it can take multiple
index_cols, which is needed for the above validations.