Adapt CSV-to-DB for foreign keys (#665)
This is a modification to the CSV-to-DB functionality that allows for 
loading of subscenario info only (without data). This allows for more 
flexibility and is helpful when foreign key implementation require 
that the "main" table (the one that matches the subscenario table name) 
in a multi-table subscenario must be loaded after another table in the 
subscenario (e.g. the inputs_temporal table must be loaded after the 
inputs_temporal_periods table). This can also be used if the 
subscenario table name does not match the inputs table name.