Remove include and required columns from CSV master (#668)
This removes the 'include' column from the CSV master file. Instead, 
we use the 'path' column to determine whether data should be loaded 
(data is loaded if a path is specified and the subscenario/table is 
skipped otherwise).

The 'required' column is also removed, as it is not used any more.